split out extraction of token from Auth Header, Url Parameter or Cookie into its own helper method. (Simplifies Module!) fixes https://github.com/dwyl/hapi-auth-jwt2/issues/67
