+++ lib/index.js
@@ -1,8 +1,8 @@
-var Boom = require('boom'); // error handling https://github.com/hapijs/boom
-var Hoek = require('hoek'); // hapi utilities https://github.com/hapijs/hoek
-var JWT  = require('jsonwebtoken'); // https://github.com/docdis/learn-json-web-tokens
-var COOKIE = require('cookie'); // highly popular decoupled cookie parser
-var pkg  = require('../package.json');
+var Boom      = require('boom'); // error handling https://github.com/hapijs/boom
+var Hoek      = require('hoek'); // hapi utilities https://github.com/hapijs/hoek
+var JWT       = require('jsonwebtoken'); // https://github.com/docdis/learn-json-web-tokens
+var extract   = require('./extract');    // extract token from Auth Header, URL or Coookie
+var pkg       = require('../package.json');
 var internals = {}; // Declare internals >> see: http://hapijs.com/styleguide
 
 exports.register = function (server, options, next) {
@@ -19,38 +19,6 @@ internals.isFunction = function (functionToCheck) {
  return functionToCheck && getType.toString.call(functionToCheck) === '[object Function]';
 };
 
-function basicChecks (token, request, reply) {
-  // rudimentary check for JWT validity see: http://git.io/xPBn for JWT format
-  if (token.split('.').length !== 3) {
-    return reply(Boom.unauthorized('Invalid token format', 'Token'));
-  }
-  else {
-    return token;
-  }
-}
-
-function extractToken (request, reply) {
-  var auth;
-  if(request.query.token) { // tokens via url: https://github.com/dwyl/hapi-auth-jwt2/issues/19
-    auth = request.query.token;
-  } // JWT tokens in cookie: https://github.com/dwyl/hapi-auth-jwt2/issues/55
-  else if (request.headers.authorization) {
-    auth = request.headers.authorization;
-  }
-  else if (request.headers.cookie) {
-    auth = COOKIE.parse(request.headers.cookie).token;
-  }
-  if (!auth && (request.auth.mode === 'optional' || request.auth.mode === 'try')) {
-    return reply.continue({ credentials: {} });
-  }
-  if (!auth) {
-    return reply(Boom.unauthorized('Missing auth token'));
-  }
-  // strip pointless "Bearer " label & any whitespace > http://git.io/xP4F
-  var token = auth.replace(/Bearer/gi,'').replace(/ /g,'');
-  return token;
-}
-
 internals.implementation = function (server, options) {
   Hoek.assert(options, 'options are required for jwt auth scheme'); // pre-auth checks
   Hoek.assert(options.key, 'options must contain secret key or key lookup function'); // no signing key
@@ -58,8 +26,7 @@ internals.implementation = function (server, options) {
 
   var scheme = {
     authenticate: function (request, reply) {
-      var token = extractToken(request, reply);
-      basicChecks(token, request, reply);
+      var token = extract(request, reply);
       var keyFunc = (internals.isFunction(options.key)) ? options.key : function (decoded, callback) { callback(null, options.key); };
       keyFunc(JWT.decode(token), function (err, key, extraInfo) {
         if (err) {
@@ -70,14 +37,11 @@ internals.implementation = function (server, options) {
         }
         var verifyOptions = options.verifyOptions || {};
         JWT.verify(token, key, verifyOptions, function (err, decoded) {
-          if (err) { // for 'try' mode we need to pass back the decoded token even if verification failed
-            var credentials = JWT.decode(token);
-            if (err.name === 'TokenExpiredError') {
-              return reply(Boom.unauthorized('Token expired', 'Token'), null, { credentials: credentials });
-            }
-            else {
-              return reply(Boom.unauthorized('Invalid token', 'Token'), null, { credentials: credentials });
-            }
+          if (err && err.name === 'TokenExpiredError') {
+            return reply(Boom.unauthorized('Token expired', 'Token'), null, { credentials: null });
+          }
+          else if (err) {
+            return reply(Boom.unauthorized('Invalid token', 'Token'), null, { credentials: null });
           }
           else { // see: http://hapijs.com/tutorials/auth for validateFunc signature
             options.validateFunc(decoded, request, function (err, valid, credentials) { // bring your own checks
